BUG32120659: Prepared statements w/o parameters violates MySQL protocol
Connector python violates MySQL protocol sending unrequired extra bytes
with a prepared statement without any parameter.
like for:
 cur.execute("SELECT 1")
Before:
  b'\x0b\x00\x00\x00\x17\x01\x00\x00\x00\x00\x01\x00\x00\x00\x01'
Now:
  b'\x0a\x00\x00\x00\x17\x01\x00\x00\x00\x00\x01\x00\x00\x00'

+++ src/mysql_capi.c
@@ -2852,7 +2852,7 @@ MySQL_stmt_prepare(MySQL *self, PyObject *args)
     MYSQL_RES *mysql_res= NULL;
     int res= 0;
     const char *stmt_char= NULL;
-    unsigned int stmt_length= 0;
+    unsigned long stmt_length= 0;
     unsigned long param_count= 0;
     PyObject *stmt;
     PyObject *prep_stmt;
@@ -2865,7 +2865,6 @@ MySQL_stmt_prepare(MySQL *self, PyObject *args)
     }
     stmt_char= PyBytes_AsString(stmt);
     stmt_length= strlen(stmt_char);
-
     mysql= &self->session;
 
     Py_BEGIN_ALLOW_THREADS
